ROX-21945: compliance results use separate pipelines for V1 and V2 (#9527)
Moving compliance check results to use separate pipelines for V1 and V2
