+++ central/sensor/service/pipeline/complianceoperatorresults/pipeline_test.go
@@ -4,9 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/gogo/protobuf/types"
 	v1ResultMocks "github.com/stackrox/rox/central/complianceoperator/checkresults/datastore/mocks"
-	v2ResultMocks "github.com/stackrox/rox/central/complianceoperator/v2/checkresults/datastore/mocks"
 	"github.com/stackrox/rox/generated/internalapi/central"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/features"
@@ -17,15 +15,12 @@ import (
 )
 
 const (
-	mockScanName      = "mock-scan"
 	mockCheckRuleName = "mock-rule"
-	mockSuiteName     = "mock-suite"
 )
 
 var (
-	createdTime = types.TimestampNow()
-	id          = uuid.NewV4().String()
-	checkID     = uuid.NewV4().String()
+	id      = uuid.NewV4().String()
+	checkID = uuid.NewV4().String()
 )
 
 func TestPipeline(t *testing.T) {
@@ -36,7 +31,6 @@ type PipelineTestSuite struct {
 	suite.Suite
 
 	v1ResultDS *v1ResultMocks.MockDataStore
-	v2ResultDS *v2ResultMocks.MockDataStore
 	mockCtrl   *gomock.Controller
 }
 
@@ -51,98 +45,17 @@ func (suite *PipelineTestSuite) SetupSuite() {
 func (suite *PipelineTestSuite) SetupTest() {
 	suite.mockCtrl = gomock.NewController(suite.T())
 	suite.v1ResultDS = v1ResultMocks.NewMockDataStore(suite.mockCtrl)
-	suite.v2ResultDS = v2ResultMocks.NewMockDataStore(suite.mockCtrl)
 }
 
 func (suite *PipelineTestSuite) TearDownTest() {
 	suite.mockCtrl.Finish()
 }
 
-func (suite *PipelineTestSuite) TestRunCreate() {
-	ctx := context.Background()
-
-	suite.v1ResultDS.EXPECT().Upsert(ctx, getV1TestRec(fixtureconsts.Cluster1)).Return(nil).Times(1)
-	suite.v2ResultDS.EXPECT().UpsertResult(ctx, getTestRec(fixtureconsts.Cluster1)).Return(nil).Times(1)
-	pipeline := NewPipeline(suite.v1ResultDS, suite.v2ResultDS)
-
-	msg := &central.MsgFromSensor{
-		Msg: &central.MsgFromSensor_Event{
-			Event: &central.SensorEvent{
-				Id:     id,
-				Action: central.ResourceAction_CREATE_RESOURCE,
-				Resource: &central.SensorEvent_ComplianceOperatorResultV2{
-					ComplianceOperatorResultV2: &central.ComplianceOperatorCheckResultV2{
-						Id:           id,
-						CheckId:      checkID,
-						CheckName:    mockCheckRuleName,
-						ClusterId:    fixtureconsts.Cluster1,
-						Status:       central.ComplianceOperatorCheckResultV2_FAIL,
-						Severity:     central.ComplianceOperatorRuleSeverity_HIGH_RULE_SEVERITY,
-						Description:  "this is a test",
-						Instructions: "this is a test",
-						Labels:       nil,
-						Annotations:  nil,
-						CreatedTime:  createdTime,
-						ScanName:     mockScanName,
-						SuiteName:    mockSuiteName,
-						Rationale:    "test rationale",
-						ValuesUsed:   []string{"var1", "var2"},
-						Warnings:     []string{"warning1", "warning2"},
-					},
-				},
-			},
-		},
-	}
-
-	err := pipeline.Run(ctx, fixtureconsts.Cluster1, msg, nil)
-	suite.NoError(err)
-}
-
-func (suite *PipelineTestSuite) TestRunDelete() {
-	ctx := context.Background()
-
-	suite.v1ResultDS.EXPECT().Delete(ctx, id).Return(nil).Times(1)
-	suite.v2ResultDS.EXPECT().DeleteResult(ctx, id).Return(nil).Times(1)
-	pipeline := NewPipeline(suite.v1ResultDS, suite.v2ResultDS)
-
-	msg := &central.MsgFromSensor{
-		Msg: &central.MsgFromSensor_Event{
-			Event: &central.SensorEvent{
-				Id:     id,
-				Action: central.ResourceAction_REMOVE_RESOURCE,
-				Resource: &central.SensorEvent_ComplianceOperatorResultV2{
-					ComplianceOperatorResultV2: &central.ComplianceOperatorCheckResultV2{
-						Id:           id,
-						CheckId:      checkID,
-						CheckName:    mockCheckRuleName,
-						ClusterId:    fixtureconsts.Cluster1,
-						Status:       central.ComplianceOperatorCheckResultV2_FAIL,
-						Severity:     central.ComplianceOperatorRuleSeverity_HIGH_RULE_SEVERITY,
-						Description:  "this is a test",
-						Instructions: "this is a test",
-						Labels:       nil,
-						Annotations:  nil,
-						CreatedTime:  createdTime,
-						ScanName:     mockScanName,
-						SuiteName:    mockSuiteName,
-						Rationale:    "test rationale",
-						ValuesUsed:   []string{"var1", "var2"},
-						Warnings:     []string{"warning1", "warning2"},
-					},
-				},
-			},
-		},
-	}
-
-	err := pipeline.Run(ctx, fixtureconsts.Cluster1, msg, nil)
-	suite.NoError(err)
-}
-
 func (suite *PipelineTestSuite) TestRunV1Create() {
 	ctx := context.Background()
 
 	suite.v1ResultDS.EXPECT().Upsert(ctx, getV1TestRec(fixtureconsts.Cluster1)).Return(nil).Times(1)
-	pipeline := NewPipeline(suite.v1ResultDS, suite.v2ResultDS)
+	pipeline := NewPipeline(suite.v1ResultDS)
 
 	msg := &central.MsgFromSensor{
 		Msg: &central.MsgFromSensor_Event{
@@ -174,7 +87,7 @@ func (suite *PipelineTestSuite) TestRunV1Delete() {
 	ctx := context.Background()
 
 	suite.v1ResultDS.EXPECT().Delete(ctx, id).Return(nil).Times(1)
-	pipeline := NewPipeline(suite.v1ResultDS, suite.v2ResultDS)
+	pipeline := NewPipeline(suite.v1ResultDS)
 
 	msg := &central.MsgFromSensor{
 		Msg: &central.MsgFromSensor_Event{
@@ -200,27 +113,6 @@ func (suite *PipelineTestSuite) TestRunV1Delete() {
 
 	err := pipeline.Run(ctx, fixtureconsts.Cluster1, msg, nil)
 	suite.NoError(err)
-}
-
-func getTestRec(clusterID string) *storage.ComplianceOperatorCheckResultV2 {
-	return &storage.ComplianceOperatorCheckResultV2{
-		Id:             id,
-		CheckId:        checkID,
-		CheckName:      mockCheckRuleName,
-		ClusterId:      clusterID,
-		Status:         storage.ComplianceOperatorCheckResultV2_FAIL,
-		Severity:       storage.RuleSeverity_HIGH_RULE_SEVERITY,
-		Description:    "this is a test",
-		Instructions:   "this is a test",
-		Labels:         nil,
-		Annotations:    nil,
-		CreatedTime:    createdTime,
-		ScanConfigName: mockSuiteName,
-		ScanName:       mockScanName,
-		Rationale:      "test rationale",
-		ValuesUsed:     []string{"var1", "var2"},
-		Warnings:       []string{"warning1", "warning2"},
-	}
 }
 
 func getV1TestRec(clusterID string) *storage.ComplianceOperatorCheckResult {
